feat : 기본적인 ui + api 연결해서 데이터 가져오기
